Refactoring all payroll functions into one folder
